backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
openai==1.82.0
python-dateutil==2.8.2 
 